requirements.txt
numpy
opencv-python
PyYAML
scikit-learn
scipy
torch
torchvision
tqdm
